+++ lib/cyrax/serializers/scope.rb
@@ -2,6 +2,7 @@ module Cyrax::Serializers
   class Scope
     def initialize(&block)
       @attrs = {}
+      @dynamic_attrs = {}
       instance_eval(&block) if block_given?
     end
 
@@ -15,19 +16,25 @@ module Cyrax::Serializers
       end
     end
 
-    def attribute(attribute, options = {})
-      @attrs ||= {}
-      @attrs[attribute] ||= attribute
+    def attribute(attribute, options = {}, &block)
+      if block_given?
+        @dynamic_attrs[attribute] = block
+      else
+        @attrs[attribute] = attribute
+      end
     end
 
-    def serialize(object)
+    def serialize(resource, options = {})
       result = {}
       @attrs.map do |attribute, options|
-        value = object.send(attribute)
+        value = resource.send(attribute)
         if options.is_a?(Cyrax::Serializers::Scope)
           value = options.serialize(value)
         end
         result[attribute] = value
+      end
+      @dynamic_attrs.map do |attribute, block|
+        result[attribute] = options[:serializer].instance_eval(&block)
       end
       result
     end